refactor(M9A): 将 M9AConfigBuilder 集成到 AutoProxyTask
- 移除 config_builder.py 文件
- 将 M9AConfigBuilder 的所有方法移入 AutoProxyTask 类：
  - build_config
  - _build_option_list
  - _build_option_list_from_user
  - _build_task_item
  - _build_adb_device_config
  - _build_ldplayer_config
  - _build_mumu_config
- 移除 config_builder 导入和初始化代码
- 更新 build_config 调用方式
